[auth] using auth middleware + random 20byte token

+++ internal/pkg/controllers/usersController.go
@@ -21,10 +21,11 @@ type SingUpRequest struct {
 }
 
 type SignUpResponse struct {
-	Id int `json:"id"`
+	Id int64 `json:"id"`
 }
 
 func SignUp(res http.ResponseWriter, req *http.Request) {
+	// TODO(): проверка куки
 	fmt.Println("createUser")
 
 	body, err := ioutil.ReadAll(req.Body)
